requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
werkzeug==2.0.1
python-dotenv==0.19.0
mysqlclient==2.2.0
pymysql==1.1.0
reportlab==4.0.4
matplotlib==3.7.1
numpy==1.24.3 